refactor(customfields): craft-static cleanup — file split + long-func extraction
craft static (diff-scoped) flagged two MAJOR findings on this arc's own
new code: customfields_http_integration_test.go crossed the repo's
500-LOC file cap, and cmd/api's run() crossed the 80-body-line long-func
budget after the schema-pool wiring landed. Neither is a BLOCKER, but
both are fixed rather than left for the pre-push gate to flag:

- Split the HTTP suite by concept: the test entry point, wire types, and
  shared createCustomField helper stay in
  customfields_http_integration_test.go; every per-scenario assertion
  helper moves to the new customfields_http_assertions_test.go. The
  remaining three inline t.Run bodies (validation, structural refusal,
  duplicate slug, unknown-id retire, invalid list object, not-picklist
  options) are extracted to named helpers alongside their siblings, so
  TestCustomFieldsHTTP itself reads as a flat list of scenario names.
- Extracted baseComposeOptions(ctx, cfg, pool, stdout) out of cmd/api's
  run(): the blobstore/keyvault/schema-pool option assembly that doesn't
  depend on the inline relay's lifecycle, so run() stays inside the
  file's long-func budget.

craft static now reports 0 blocker/major/minor over every file this arc
touched; make check and the customfields HTTP/store integration lanes
stay green.
